feat: Added support for subtitles (Movies only)
