+++ android/app/src/main/res/values/colors.xml
@@ -3,4 +3,6 @@
     <color name="background">#111214</color>
     <color name="background_secondary">#0D0E0F</color>
     <color name="dropdown_background">#212225</color>
+
+    <color name="text_color">#FAFAFA</color>
 </resources>
