+++ android/app/src/main/res/values/styles.xml
@@ -1,9 +1,11 @@
 <resources>
 
     <!-- Base application theme. -->
-    <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
-        <!-- Customize your theme here. -->
+    <style name="AppTheme" parent="Theme.AppCompat.NoActionBar">
+        <!--<item name="colorPrimary">@color/background</item>-->
+        <item name="colorAccent">@color/text_color</item>
 
+        <item name="android:windowBackground">@color/background</item>
         <item name="android:spinnerItemStyle">@style/SpinnerItem</item>
         <item name="android:spinnerDropDownItemStyle">@style/SpinnerDropDownItem</item>
         <item name="android:navigationBarColor">@color/background_secondary</item>
@@ -19,17 +21,18 @@
     <style name="SplashTheme" parent="Theme.AppCompat.Light.NoActionBar">
         <item name="android:statusBarColor">@color/background</item>
         <item name="android:navigationBarColor">@color/background</item>
+        <item name="android:windowBackground">@color/background</item>
     </style>
 
     <style name="SpinnerItem" parent="Theme.AppCompat.Light.NoActionBar">>
-        <item name="android:textColor">#FAFAFA</item>
+        <item name="android:textColor">@color/text_color</item>
         <item name="android:textSize">18dp</item>
         <item name="android:fontFamily">Roboto-Bold</item>
         <item name="android:gravity">center</item>
     </style>
 
     <style name="SpinnerDropDownItem" parent="Theme.AppCompat.Light.NoActionBar">>
-        <item name="android:textColor">#FAFAFA</item>
+        <item name="android:textColor">@color/text_color</item>
         <item name="android:textSize">18dp</item>
         <item name="android:fontFamily">Roboto-Bold</item>
         <item name="android:background">@color/dropdown_background</item>
